50-Project/52_FunnelDesignSkill/00_brand-extraction/scripts/requirements.txt
# Brand Extraction Dependencies

requests>=2.28.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
